+++ FamilyFinance/Services/Interfaces/IDirectaImportService.cs
@@ -1,0 +1,8 @@
+using FamilyFinance.Models.Import;
+
+namespace FamilyFinance.Services;
+
+public interface IDirectaImportService
+{
+    Task<DirectaImportResult> ParsePortfolioCsvAsync(Stream content);
+}